generator for scope
