+++ tests/models/_types.ts
@@ -24,8 +24,9 @@ declare module "accel-record" {
   function meta<T>(model: T): Meta<T>;
 
   interface Relation<T, M> {
-    john: (M extends UserMeta ? typeof UserModel['john'] : never)
-    adults: (M extends UserMeta ? typeof UserModel['adults'] : never)
+    john: (T extends User ? typeof UserModel['john'] : never);
+    adults: (T extends User ? typeof UserModel['adults'] : never);
+    teenagers: (T extends User ? typeof UserModel['teenagers'] : never);
   }
 }
